move SyscallN test
windows doesn't support dlysym
