+++ syscall_test.go
@@ -6,9 +6,6 @@ package purego_test
 import (
 	"os"
 	"testing"
-	"unsafe"
-
-	"github.com/ebitengine/purego"
 )
 
 func TestOS(t *testing.T) {
@@ -20,17 +17,5 @@ func TestOS(t *testing.T) {
 	err = os.Unsetenv("TESTING")
 	if err != nil {
 		t.Errorf("failed to Unsetenv: %s", err)
-	}
-}
-
-func TestSyscallN(t *testing.T) {
-	var dlsym uintptr
-	var err error
-	if dlsym, err = purego.Dlsym(purego.RTLD_DEFAULT, "dlsym"); err != nil {
-		t.Errorf("Dlsym with RTLD_DEFAULT failed: %v", err)
-	}
-	r1, _, err2 := purego.SyscallN(dlsym, purego.RTLD_DEFAULT, uintptr(unsafe.Pointer(&[]byte("dlsym\x00")[0])))
-	if dlsym != r1 {
-		t.Fatalf("SyscallN didn't return the same result as purego.Dlsym: %d", err2)
 	}
 }
